Make initialization more sensible
separate parameters existed for gene and synomyms, but the only thing
that is done is adding all of their tokens to the list of
stopwords. It seems more reasonable to have the variable name follow
function rather than intended use. Attributes that can be set by other
methods have been initialized with value None